requirements.txt
Django>=4.2.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
